+++ CHANGES.md
@@ -49,6 +49,7 @@ Changes in 5.0
 - Save Template even if no default template found in Template.fnc.php
 - Fix $staff_id for admin user in \_makeLetterGrade.fnc.php
 - Add Min. and Max. Grades option in ReportCards.fnc.php
+- JS remove deprecated IE8 HTML5 tags fix & touchScroll() in warehouse.js
 
 Changes in 4.9.6
 ----------------